Home: change nav label, title, icon; improve/break out panel title
