+++ src/components/panels/PanelTitleBar.tsx
@@ -1,26 +1,14 @@
 import React, { FC, useEffect } from 'react'
-import {
-  useLocation,
-  Route,
-  Switch,
-  Link as RouterLink,
-} from 'react-router-dom'
+import { useLocation, Route, Switch } from 'react-router-dom'
 import { makeStyles, Theme, createStyles } from '@material-ui/core/styles'
-import {
-  AppBar,
-  IconButton,
-  Toolbar,
-  Typography,
-  Tooltip,
-} from '@material-ui/core'
-import { GoSearch } from 'react-icons/go'
-import { MdClose } from 'react-icons/md'
+import { AppBar, IconButton, Toolbar, Tooltip } from '@material-ui/core'
+import { HiOutlineSearch } from 'react-icons/hi'
+import { CgClose } from 'react-icons/cg'
 
 import { usePanelDispatch } from 'components/panels'
-import { icons } from 'components/config'
 import { routes } from 'components/config/api'
-import { pluralize } from 'components/explore/utils'
 import { SplitCrumbs } from './SplitCrumbs'
+import { PanelTitleRoutes } from './PanelTitleRoutes'
 
 type StyleProps = { hide?: boolean }
 
@@ -40,30 +28,12 @@ const useStyles = makeStyles((theme: Theme) =>
       maxHeight: (props: StyleProps) => (props.hide ? 0 : 48),
       zIndex: (props: StyleProps) => (props.hide ? -1 : 1),
     },
-    colorDefault: {
-      backgroundColor: 'transparent',
-    },
     toolbar: {
       backgroundColor: theme.palette.primary.dark,
       borderTopLeftRadius,
       borderTopRightRadius,
       padding: '0 0.5rem',
       justifyContent: 'space-between',
-    },
-    panelTitleAndIcon: {
-      display: 'flex',
-      alignItems: 'center',
-      '& > svg': {
-        color: theme.palette.text.hint,
-        marginRight: 4,
-        fontSize: '1.25rem',
-      },
-    },
-    panelTitleText: {
-      fontSize: '1.5rem',
-      [theme.breakpoints.down('sm')]: {
-        fontSize: '1.25rem',
-      },
     },
     rightSideBtns: {
       '& > * + *': {
@@ -73,100 +43,32 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 )
 
-type PanelTitleProps = {
-  text: string
-  icon?: React.ReactNode
-}
+const PanelCloseBtn: FC = (props) => {
+  const { pathname } = useLocation()
+  const panelDispatch = usePanelDispatch()
 
-// TODO: move this and the big Routes thing into new component
-const PanelTitle: FC<PanelTitleProps> = (props) => {
-  const { text, icon } = props
-  const classes = useStyles({})
+  useEffect(() => {
+    panelDispatch({ type: 'TOGGLE_SEARCH_TABS', payload: false })
+  }, [panelDispatch, pathname])
 
   return (
-    <div className={classes.panelTitleAndIcon}>
-      {icon}
-      <Typography
-        variant="h6"
-        component="h2"
-        className={classes.panelTitleText}
-      >
-        {text}
-      </Typography>
-    </div>
-  )
-}
-
-const LinkToHomeBtn: FC = (props) => {
-  return (
-    <Tooltip title="Go to Home panel">
+    <Tooltip title="Close panel">
       <IconButton
         size="small"
-        aria-label="go home"
+        aria-label="panel close"
         color="inherit"
-        to="/"
-        component={RouterLink}
+        onClick={() =>
+          panelDispatch({ type: 'TOGGLE_MAIN_PANEL', payload: false })
+        }
       >
-        {icons.Home}
+        <CgClose />
       </IconButton>
     </Tooltip>
   )
 }
 
-const PanelTitleRoutes: FC<{ panelTitle: string }> = (props) => {
-  const { panelTitle } = props
-
-  return (
-    <Switch>
-      <Route path="/" exact>
-        {/* Flex spacer */}
-        <div />
-        <PanelTitle text="Search and Display Sites" />
-      </Route>
-      <Route path={routes.none}>
-        <PanelTitle text="No site selected" />
-      </Route>
-      <Route path={routes.data}>
-        <PanelTitle text=" " />
-      </Route>
-      <Route
-        path={['/Explore/:field/:value/:language/:id', routes.details]}
-        exact
-      >
-        <PanelTitle text="Site Details" icon={icons.SiteDetails} />
-      </Route>
-      <Route
-        path={['/Explore/:field/:value/:language', routes.languageInstance]}
-        exact
-      >
-        <PanelTitle text="Language" icon={icons.Language} />
-      </Route>
-      <Route path="/Explore/:field" exact>
-        <PanelTitle text={pluralize(panelTitle)} icon={icons[panelTitle]} />
-      </Route>
-      <Route path="/Explore/:field/:value" exact>
-        <PanelTitle text={panelTitle} icon={icons[panelTitle]} />
-      </Route>
-      <Route path="/Census">
-        {/* Census just needs panel heading override */}
-        <LinkToHomeBtn />
-        <PanelTitle text="Census Language Data" icon={icons[panelTitle]} />
-      </Route>
-      <Route path="/:level1" exact>
-        {/* Home btn on /TopLevelRoutes looks balanced on left */}
-        <LinkToHomeBtn />
-        <PanelTitle text={panelTitle} icon={icons[panelTitle]} />
-      </Route>
-    </Switch>
-  )
-}
-
-export const PanelTitleBar: FC<{ hide: boolean }> = (props) => {
-  const { hide } = props
-  const classes = useStyles({ hide })
+const ToggleSearchMenuBtn: FC = () => {
   const { pathname } = useLocation()
-  // Lil' gross, but use Level2 route name first, otherwise Level1
-  const panelTitle = pathname.split('/')[2] || pathname.split('/')[1]
   const panelDispatch = usePanelDispatch()
 
   const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
@@ -177,32 +79,33 @@ export const PanelTitleBar: FC<{ hide: boolean }> = (props) => {
     panelDispatch({ type: 'TOGGLE_SEARCH_TABS', payload: false })
   }, [panelDispatch, pathname])
 
+  return (
+    <Tooltip title="Toggle search menu">
+      <IconButton
+        size="small"
+        aria-label="toggle search menu"
+        color="inherit"
+        onClick={handleClick}
+      >
+        <HiOutlineSearch />
+      </IconButton>
+    </Tooltip>
+  )
+}
+
+export const PanelTitleBar: FC<{ hide: boolean }> = (props) => {
+  const { hide } = props
+  const classes = useStyles({ hide })
+  const { pathname } = useLocation()
+  // Lil' gross, but use Level2 route name first, otherwise Level1
+  const panelTitle = pathname.split('/')[2] || pathname.split('/')[1]
+
   const RightSideBtns = (
     <div className={classes.rightSideBtns}>
       <Route path="/Explore">
-        <Tooltip title="Toggle search menu">
-          <IconButton
-            size="small"
-            aria-label="toggle search menu"
-            color="inherit"
-            onClick={handleClick}
-          >
-            <GoSearch />
-          </IconButton>
-        </Tooltip>
+        <ToggleSearchMenuBtn />
       </Route>
-      <Tooltip title="Close panel">
-        <IconButton
-          size="small"
-          aria-label="panel close"
-          color="inherit"
-          onClick={() =>
-            panelDispatch({ type: 'TOGGLE_MAIN_PANEL', payload: false })
-          }
-        >
-          <MdClose />
-        </IconButton>
-      </Tooltip>
+      <PanelCloseBtn />
     </div>
   )
